bmv-backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
httpx
python-dotenv
python-jose[cryptography]
supabase
pytest
